Separate files and rename functions
